Completed #80

+++ neon/Service/DataGateway.swift
@@ -11,13 +11,13 @@ import CloudKit
 import Firebase
 
 class DataGateway {
-    
+	
     static let shared = DataGateway()
-    
+	
     func fetchAgendaItems(completion: @escaping (_ todaysAgendaItems: [Int: AgendaItem], _ tomorrowsAgendaItems: [Int: AgendaItem], _ success: Bool) -> ()) {
         var todaysAgendaItems = [Int: AgendaItem]()
         var tomorrowsAgendaItems = [Int: AgendaItem]()
-        
+		
         if CalendarGateway.shared.hasPermission() {
             for event in CalendarGateway.shared.importTodaysEvents() {
                 for i in event.startTime...event.endTime {
@@ -26,7 +26,7 @@ class DataGateway {
                     todaysAgendaItems[i] = agendaItem
                 }
             }
-            
+			
             for event in CalendarGateway.shared.importTomorrowsEvents() {
                 for i in event.startTime...event.endTime {
                     var agendaItem = AgendaItem(title: event.title)
@@ -35,10 +35,10 @@ class DataGateway {
                 }
             }
         }
-        
+		
         let database = CKContainer.default().privateCloudDatabase
         let query = CKQuery(recordType: "AgendaRecord", predicate: NSPredicate(value: true))
-        
+		
         database.perform(query, inZoneWith: nil) { (records, error) in
             if error == nil {
                 records?.forEach({ (record) in
@@ -46,7 +46,7 @@ class DataGateway {
                     guard let title = record.value(forKey: "title") as? String else { return }
                     guard let hour = record.value(forKey: "hour") as? Int else { return }
                     guard let day = record.value(forKey: "day") as? Date else { return }
-                    
+					
                     // Only pull the tasks that are in today and aren't already on device
                     if Calendar.current.isDateInToday(day) {
                         todaysAgendaItems[hour] = AgendaItem(with: id, and: title)
@@ -54,7 +54,7 @@ class DataGateway {
                         tomorrowsAgendaItems[hour] = AgendaItem(with: id, and: title)
                     }
                 })
-                
+				
                 completion(todaysAgendaItems, tomorrowsAgendaItems, true)
             } else {
 				AnalyticsGateway.shared.logCloudError(for: error!.localizedDescription)
@@ -62,16 +62,16 @@ class DataGateway {
             }
         }
     }
-    
+	
     func save(_ agendaItem: AgendaItem, for hour: Int, today: Bool) {
         let database = CKContainer.default().privateCloudDatabase
-        
+		
         let record = CKRecord(recordType: "AgendaRecord")
         record.setObject((today ? Date() : Calendar.current.date(byAdding: .day, value: 1, to: Date())!) as CKRecordValue, forKey: "day")
         record.setObject(agendaItem.id as CKRecordValue, forKey: "id")
         record.setObject(agendaItem.title as CKRecordValue, forKey: "title")
         record.setObject(hour as CKRecordValue, forKey: "hour")
-        
+		
         database.save(record) { (record, error) in
             if error == nil {
 				self.logHourBlock(with: agendaItem.title)
@@ -79,11 +79,11 @@ class DataGateway {
 			}
         }
     }
-    
+	
     func delete(_ agendaItem: AgendaItem) {
         let database = CKContainer.default().privateCloudDatabase
         let query = CKQuery(recordType: "AgendaRecord", predicate: NSPredicate(value: true))
-        
+		
         database.perform(query, inZoneWith: nil) { (records, error) in
             records?.forEach({ (record) in
                 guard let id = record.value(forKey: "id") as? String else { return }
@@ -93,38 +93,42 @@ class DataGateway {
             })
         }
     }
-    
+	
     func deletePastAgendaRecords() {
         let database = CKContainer.default().privateCloudDatabase
         let query = CKQuery(recordType: "AgendaRecord", predicate: NSPredicate(value: true))
-        
+		
         database.perform(query, inZoneWith: nil) { (records, error) in
-            records?.forEach({ (record) in
-                guard let day = record.value(forKey: "day") as? Date else { return }
-                
-                if (!Calendar.current.isDateInToday(day) && !Calendar.current.isDateInTomorrow(day)) {
-                    database.delete(withRecordID: record.recordID, completionHandler: { (recordID, error) in })
-                }
-            })
+			if error == nil {
+            	records?.forEach({ (record) in
+                	guard let day = record.value(forKey: "day") as? Date else { return }
+					
+                	if (!Calendar.current.isDateInToday(day) && !Calendar.current.isDateInTomorrow(day)) {
+                    	database.delete(withRecordID: record.recordID, completionHandler: { (recordID, error) in })
+                	}
+				})
+			} else {
+				AnalyticsGateway.shared.logCloudError(for: error!.localizedDescription)
+			}
         }
     }
-    
+	
     func getTotalAgendaCount() -> Int {
         guard let totalAgendaCount = UserDefaults.standard.object(forKey: "totalAgendaCount") as? Int else {
             return 0
         }
         return totalAgendaCount
     }
-    
+	
     private func incrementTotalAgendaCount() {
         let totalAgendaCount = UserDefaults.standard.object(forKey: "totalAgendaCount") as? Int
-        
+		
         if totalAgendaCount == nil {
             UserDefaults.standard.set(1, forKey: "totalAgendaCount")
         } else {
             UserDefaults.standard.set(totalAgendaCount! + 1, forKey: "totalAgendaCount")
         }
-        
+		
         UserDefaults.standard.synchronize()
     }
